requirements.txt
omegaconf==2.1.0.dev26
hydra-core==1.1.0.dev6
ray==1.2.0
cloudpickle==1.6.0
pickle5==0.0.11
hydra-ray-launcher==1.1.0.dev1
torch==1.7.1
torchvision==0.8.2
torchtext
pandas==1.2.1
tabulate==0.8.7
s3fs==0.4.2
tqdm==4.56.0
tensorboardX==2.1
seaborn==0.11.1
git+https://github.com/samuelstanton/upcycle.git@master
